Backend/requirements.txt

Django==4.1.7

django-cors-headers==3.14.0
django-treebeard==4.7
djangorestframework==3.14.0

mysql-client==0.0.1

requests==2.31.0

sqlparse==0.4.3

